[frontend] Convert Taxii to tsx
